+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'Portfolio1'
+rootProject.name = 'CommunityEventsApp'
