+++ src/main/java/com/codestep/CommunityEventsApp/CommunityEventsAppApplication.java
@@ -1,0 +1,13 @@
+package com.codestep.CommunityEventsApp;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CommunityEventsAppApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(CommunityEventsAppApplication.class, args);
+	}
+
+}
