+++ src/main/java/com/codestep/CommunityEventsApp/Exceptions/ValidationException.java
@@ -1,0 +1,16 @@
+package com.codestep.CommunityEventsApp.Exceptions;
+
+import org.springframework.validation.BindingResult;
+
+public class ValidationException extends RuntimeException {
+
+    private final BindingResult result;
+
+    public ValidationException(BindingResult result) {
+        this.result = result;
+    }
+
+    public BindingResult getBindingResult() {
+        return result;
+    }
+}
