+++ src/main/java/com/codestep/CommunityEventsApp/Validators/UserExistingValidator.java
@@ -1,0 +1,26 @@
+package com.codestep.CommunityEventsApp.Validators;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.provisioning.UserDetailsManager;
+
+import com.codestep.CommunityEventsApp.Annotations.UserExisting;
+
+import jakarta.validation.ConstraintValidator;
+import jakarta.validation.ConstraintValidatorContext;
+
+public class UserExistingValidator implements ConstraintValidator<UserExisting, String>{
+	@Autowired
+	UserDetailsManager userDetailsManager;
+	
+	@Override
+	public void initialize(UserExisting userExisting) {
+	}
+	
+	@Override
+	public boolean isValid(String input, ConstraintValidatorContext cxt) {
+		if(input == null) {
+			return true;
+		}
+		return !userDetailsManager.userExists(input);
+	}
+}
