--- src/main/java/com/codestep/portfolio1app/Portfolio1Application.java
@@ -1,13 +1,0 @@
-package com.codestep.portfolio1app;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Portfolio1Application {
-
-	public static void main(String[] args) {
-		SpringApplication.run(Portfolio1Application.class, args);
-	}
-
-}
