--- src/main/java/com/codestep/portfolio1app/repositories/PrefectureRepository.java
@@ -1,10 +1,0 @@
-package com.codestep.portfolio1app.repositories;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import com.codestep.portfolio1app.entities.Prefecture;
-
-@Repository
-public interface PrefectureRepository extends JpaRepository<Prefecture, Long>{
-}
