+++ src/test/java/com/codestep/CommunityEventsApp/CommunityEventsAppApplicationTests.java
@@ -1,0 +1,13 @@
+package com.codestep.CommunityEventsApp;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CommunityEventsAppApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
