--- src/test/java/com/codestep/portfolio1app/Portfolio1ApplicationTests.java
@@ -1,13 +1,0 @@
-package com.codestep.portfolio1app;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class Portfolio1ApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
